requirements.txt
annotated-types==0.7.0
anyio==4.9.0
backports.tarfile==1.2.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
click-log==0.4.0
colorama==0.4.6
cryptography==44.0.3
deprecation==2.1.0
distro==1.9.0
docutils==0.21.2
dotenv==0.9.9
dotty-dict==1.3.1
faiss-cpu==1.11.0
filelock==3.18.0
Flask==2.3.2
Flask-Cors==4.0.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
gotrue==1.3.1
gunicorn==21.2.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.16.4
idna==3.10
importlib_metadata==8.7.0
invoke==1.7.3
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.0
keyring==25.6.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.4.2
nh3==0.2.21
nltk==3.9.1
numpy==1.26.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
# Let PyTorch manage CUDA dependencies automatically
# nvidia-cudnn-cu12 - managed by torch
# nvidia-cufft-cu12 - managed by torch  
# nvidia-curand-cu12 - managed by torch
# nvidia-cusolver-cu12 - managed by torch
# nvidia-cusparse-cu12 - managed by torch
# nvidia-nccl-cu12 - managed by torch
# nvidia-nvjitlink-cu12 - managed by torch
# nvidia-nvtx-cu12 - managed by torch
packaging==24.2
Pillow==10.4.0
platformdirs==4.3.6
postgrest==0.10.8
protobuf==5.28.3
psutil==6.1.0
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
realtime==1.0.6
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
SecretStorage==3.3.3
sentence-transformers==3.3.1
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
storage3==0.8.1
supabase==2.9.1
supafunc==0.5.1
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.47.1
triton==3.1.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==2.3.6
zipp==3.21.0